# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TAU Service API",
    "description": "TAU Service Template API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's profile",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User's unique identifier"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "name": {
            "type": "string",
            "description": "User's display name"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "User's assigned roles"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Last login timestamp"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "roles",
          "preferences",
          "created_at",
          "updated_at"
        ]
      },
      "PreferencesResponse": {
        "type": "object",
        "properties": {
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark"
            ],
            "description": "User's preferred theme"
          }
        },
        "required": [
          "theme"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserResponse"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for programmatic handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationError"
                },
                "description": "Validation error details (only present for validation errors)"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Field that failed validation"
          },
          "message": {
            "type": "string",
            "description": "Validation error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
...
    }
  },
...
}